requirements.txt
# Core dependencies
numpy>=1.24.0
opencv-python>=4.8.0
pillow>=10.0.0
pyyaml>=6.0
pymavlink>=2.4.0
pyserial>=3.5
requests>=2.31.0
scipy>=1.11.0
matplotlib>=3.7.0

# Machine Learning (optional but recommended)
torch>=2.0.0
torchvision>=0.15.0
onnxruntime>=1.15.0
ultralytics>=8.0.0

# Simulation
pybullet>=3.2.5
gymnasium>=0.29.0

# Development
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0
isort>=5.12.0
